@@ -1,21 +1,21 @@
 {
 	"dependencies": {
-		"svelte": "^4.2.20"
+		"svelte": "^5.46.1"
 	},
 	"devDependencies": {
 		"@sveltejs/adapter-cloudflare": "^7.2.4",
-		"@sveltejs/adapter-node": "^5.3.3",
-		"@sveltejs/kit": "^2.46.5",
-		"@sveltejs/vite-plugin-svelte": "^3.1.2",
-		"@types/node": "^22.18.10",
-		"autoprefixer": "^10.4.21",
+		"@sveltejs/adapter-node": "^5.4.0",
+		"@sveltejs/kit": "^2.49.4",
+		"@sveltejs/vite-plugin-svelte": "^6.2.4",
+		"@types/node": "^25.0.3",
+		"autoprefixer": "^10.4.23",
 		"postcss": "^8.5.6",
-		"postcss-nested": "^6.2.0",
-		"prettier-plugin-svelte": "^3.4.0",
-		"sass-embedded": "^1.93.2",
-		"tailwindcss": "^3.4.18",
+		"postcss-nested": "^7.0.2",
+		"prettier-plugin-svelte": "^3.4.1",
+		"sass-embedded": "^1.97.2",
+		"tailwindcss": "^4.1.18",
 		"typescript": "^5.9.3",
-		"vite": "^5.4.20"
+		"vite": "^7.3.1"
 	},
 	"name": "web",
 	"private": true,
