@@ -1,25 +1,51 @@
 <script lang="ts">
-	import { page } from "$app/stores";
+	import "../app.css";
+
+	import {  onMount } from "svelte";
+
+	import { page } from "$app/state";
 
 	import Footer from "$lib/components/Footer.svelte";
 	import MobileNavbar from "$lib/components/MobileNavbar.svelte";
 	import Navbar from "$lib/components/Navbar.svelte";
 
-	import {  onMount } from "svelte";
-	import "../app.scss";
 	import { _, dictionary, getPage, getPageTagForLocalizedPageName, initI18n, locale } from "$lib/i18n";
 	import { goto } from "$app/navigation";
+	import { fetchData } from "$lib/utils/fetchData.svelte";
 
-	$: innerWidth = 0;
-	$: lang = $page.params?.locale;
+	const langFetch = $derived(fetchData<{ [key: string]: string }>(`/i18n/${$locale}`));
 
-	export let data;
-	$dictionary = data.dic;
+	$effect(() => {
+		if (langFetch.ok && langFetch.data) {
+			const dicLocale = $locale || "en";
+
+			dictionary.update(d => {
+				d[dicLocale] = {
+					...(d[dicLocale] || {}),
+					...langFetch.data!,
+				};
+				return d;
+			});
+		}
+	});
 
 	onMount(() => {
-		if ($page.status !== 200) return;
+		if (page.status !== 200) return;
+
+
+	});
+
+	onMount(() => {
+		const browserLanguage = navigator.language;
+		const locale = browserLanguage.split("-")[0];
+		
+
+		// i18n
+		initI18n("en");
+		langFetch.do();
 
 		document.documentElement.lang = $locale;
+
 
 		setTimeout(
 			() =>
